IN-196 refactor: change configuration
