+++ src/features/createPost/model/adding/index.ts
@@ -1,0 +1,2 @@
+export * from './addPhotoSchema'
+export * from './addPhotosCheck'
