+++ src/features/createPost/model/index.ts
@@ -1,6 +1,6 @@
+export * from './addPostStore'
+export * from './adding'
 export * from './constants'
+export * from './cropping'
+export * from './publication'
 export * from './types'
-export * from './useCityAutocomplete'
-export * from 'src/features/createPost/model/addPostStore'
-export * from 'src/features/createPost/model/adding/addPhotoSchema'
-export * from 'src/features/createPost/model/adding/addPhotosCheck'
