+++ src/features/createPost/model/publication/index.ts
@@ -1,0 +1,1 @@
+export * from './useCityAutocomplete'
