+++ src/features/createPost/ui/cropping/ControllersPanel/AspectControllerButtons.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react'
-import { ComponentPropsWithoutRef, ElementType } from 'react'
 
 import {
   BookPhotoOutline,
@@ -7,22 +6,20 @@ import {
   LandscapePhotoOutline,
   SquarePhotoOutline,
 } from '@/assets/icons'
-import { Typography, cn, typographyVariants } from '@/common'
-import { addPostStore } from '@/features/createPost'
+import {
+  AspectButton,
+  ScaleSizeButtonType,
+  addPostStore,
+} from '@/features/createPost'
 import { observer } from 'mobx-react-lite'
-
-type ScaleSizeButtonType = {
-  aspect: number
-  icon: ElementType
-  title: string
-}
 
 type Props = {
   id: string
 }
 
 export const AspectControllerButtons = observer(({ id }: Props) => {
-  //todo перевод
+  const changeAspect = addPostStore.changeAspect
+  const currentRatio = addPostStore.getAspect(id)
   const sizes: ScaleSizeButtonType[] = [
     {
       aspect: addPostStore.getOriginAspect(id),
@@ -34,13 +31,10 @@ export const AspectControllerButtons = observer(({ id }: Props) => {
     { aspect: 1.7, icon: LandscapePhotoOutline, title: '16:9' },
   ]
 
-  const changeAspect = addPostStore.changeAspect
-  const currentRatio = addPostStore.getAspect(id)
-
   return (
     <>
       {sizes.map((el, index) => (
-        <ScaleSizeButton
+        <AspectButton
           icon={el.icon}
           isCurrent={currentRatio === el.aspect}
           key={index}
@@ -51,46 +45,3 @@ export const AspectControllerButtons = observer(({ id }: Props) => {
     </>
   )
 })
-
-type ScaleSizeButtonPropsType = {
-  isCurrent: boolean
-} & ComponentPropsWithoutRef<'span'> &
-  Omit<ScaleSizeButtonType, 'aspect'>
-
-const ScaleSizeButton = ({
-  icon,
-  isCurrent,
-  onClick,
-  title,
-}: ScaleSizeButtonPropsType) => {
-  const Component = icon
-
-  return (
-    <span
-      className={
-        'flex justify-between gap-8 cursor-pointer hover:text-accent-500 transition'
-      }
-      onClick={onClick}
-    >
-      {isCurrent ? (
-        <span
-          className={cn(
-            typographyVariants({ variant: 'h3' }),
-            'text-light-100'
-          )}
-        >
-          {title}
-        </span>
-      ) : (
-        <Typography
-          className={'text-light-900'}
-          variant={'reg16'}
-        >
-          {title}
-        </Typography>
-      )}
-
-      <Component className={'w-6 h-6'} />
-    </span>
-  )
-}
